version2/backend/requirements.txt
openai
agents
python-dotenv
aiohttp
fastapi
uvicorn[standard]
mcp
sqlalchemy
pydantic
python-dotenv
requests
